backend/requirements.txt
pandas
numpy
matplotlib
scikit-fuzzy
scikit-learn
sentence-transformers
faiss-cpu
openai
streamlit
plotly
